face-service/pyproject.toml
[project]
name = "face-service"
version = "0.1.0"
description = "Face recognition microservice using InsightFace"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "insightface>=0.7.3",
    "onnxruntime-migraphx>=1.23.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.26.0,<2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[project.scripts]
face-service = "face_service.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/face_service"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
